- Added support for VGA cursor (see vgaterm_mvcursor()) - Changed inb/outb calls to use 16-bit port addresses - Changed kernel_main to kmain to be more in-line with POSIX variants - Added support to clear VGA terminal (see vgaterm_cls()).

+++ kernel/include/kernel/port.h
@@ -28,13 +28,13 @@ static inline void io_wait(void)
     __asm__ __volatile__ ( "outb %b0, $0x80" : : "a"(0) );
 }
 
-static inline uint8_t inb (uint8_t port) {
+static inline uint16_t inb (uint16_t port) {
 	uint8_t ret;
 	__asm__ __volatile__ ("inb %w1, %0" : "=a" (ret) : "dN" (port));
 	return ret;
 }
 
-static inline void outb (uint8_t port, uint8_t data) {
+static inline void outb (uint16_t port, uint8_t data) {
 	__asm__ __volatile__ ("outb %1, %w0" : : "dN" (port), "a" (data));
 }
